+++ src/app/student/map/editor/EditorMapComponent.tsx
@@ -1,6 +1,8 @@
 "use client"
 
 import { BoothMarker } from "@/app/student/map/_components/BoothMarker"
+import { EditorToolbar } from "@/app/student/map/editor/EditorToolbar"
+import { FeatureEditor } from "@/app/student/map/editor/FeatureEditor"
 import {
   Booth,
   BoothID,
@@ -13,8 +15,6 @@ import {
 import { Location } from "@/app/student/map/lib/locations"
 import { getPolygonCenter } from "@/app/student/map/lib/utils"
 import { Exhibitor } from "@/components/shared/hooks/api/useExhibitors"
-import { Button } from "@/components/ui/button"
-import { cn } from "@/lib/utils"
 import MapboxDraw, { DrawMode } from "@mapbox/mapbox-gl-draw"
 import { Feature, Polygon } from "geojson"
 import "maplibre-gl/dist/maplibre-gl.css"
@@ -36,14 +36,7 @@ const backgroundLayerStyle: BackgroundLayer = {
   }
 }
 
-type FeatureMap = Map<BoothID, GeoJsonBooth>
-
-function toFeatureCollection(featMap: FeatureMap) {
-  return {
-    type: "FeatureCollection",
-    features: Array.from(featMap.values())
-  }
-}
+export type FeatureMap = Map<BoothID, GeoJsonBooth>
 
 export default function EditorMapComponent({
   location,
@@ -54,19 +47,51 @@ export default function EditorMapComponent({
 }) {
   const mapRef = useRef<MapRef>(null)
 
-  const [mapLoaded, setMapLoaded] = useState(false)
-
   const [drawMode, setDrawMode] = useState<DrawMode>("simple_select")
 
-  const [activeFeatureId, setActiveFeatureId] = useState<BoothID | null>(null)
+  const [mapLoaded, setMapLoaded] = useState(false)
 
   const [showBuildings, setShowBuildings] = useState(false)
 
   const geoJsonData = showBuildings ? geoJsonBuildingData : geoJsonBoothData
-
+  
+  // this is the main data structure for storing the editable features
+  // i think the reason for using a ref here instead of state was to avoid unnecessary rerenders,
+  // and also to simplify update operations since you dont need to care about immutability like with state
   const featureMap = useRef<FeatureMap>(
     new Map(geoJsonData.features.map(feat => [feat.properties.id, feat]))
   )
+
+  const [activeFeatureId, setActiveFeatureId] = useState<BoothID | null>(null)
+
+  const activeFeature = activeFeatureId
+    ? featureMap.current.get(activeFeatureId)!
+    : null
+
+  const popupLocation = activeFeature ? getPolygonCenter(activeFeature) : null
+
+  const [booths, setBooths] = useState<Booth[]>([])
+  
+  const markers = booths.map(booth => (
+    <BoothMarker key={booth.id} booth={booth} scale={1} />
+  ))
+  
+  const draw = useMemo(() => {
+    return new MapboxDraw({
+      displayControlsDefault: false,
+      defaultMode: "simple_select"
+    })
+  }, [])
+
+  // featureMap api functions ----------------------------
+
+  function setFeature(feat: GeoJsonBooth) {
+    featureMap.current.set(feat.properties.id, feat)
+  }
+
+  function deleteFeature(feat: GeoJsonBooth) {
+    featureMap.current.delete(feat.properties.id)
+  }
 
   function resetFeatures(geoJsonData: GeoJsonBoothsData) {
     featureMap.current = new Map(
@@ -76,12 +101,7 @@ export default function EditorMapComponent({
     updateDrawFeatures()
   }
 
-  function setFeature(feat: GeoJsonBooth) {
-    featureMap.current.set(feat.properties.id, feat)
-  }
-  function deleteFeature(feat: GeoJsonBooth) {
-    featureMap.current.delete(feat.properties.id)
-  }
+  // Event handlers --------------------------------------
 
   function onModeChange(e: { mode: DrawMode }) {
     setDrawMode(e.mode)
@@ -112,20 +132,15 @@ export default function EditorMapComponent({
     e.features.forEach(feat => setFeature(feat))
     updateBooths()
   }
-
+  
   function onDelete(e: { features: GeoJsonBooth[] }) {
     e.features.forEach(feat => deleteFeature(feat))
     updateBooths()
   }
 
-  const draw = useMemo(() => {
-    return new MapboxDraw({
-      displayControlsDefault: false,
-      defaultMode: "simple_select"
-    })
-  }, [])
+  // Effects ----------------------------------------------
 
-  useEffect(() => {
+  function initializeDraw() {
     if (!mapLoaded || !mapRef.current) return
     const map = mapRef.current
 
@@ -149,7 +164,7 @@ export default function EditorMapComponent({
       map.off("draw.selectionchange", onSelectionChange)
       map.off("draw.modechange", onModeChange)
     }
-  }, [mapLoaded, location])
+  }
 
   function updateDrawFeatures() {
     if (!mapLoaded) return
@@ -162,14 +177,6 @@ export default function EditorMapComponent({
     draw.set(currentFeatures)
   }
 
-  useEffect(updateDrawFeatures, [location, mapLoaded])
-
-  const [booths, setBooths] = useState<Booth[]>([])
-
-  const markers = booths.map(booth => (
-    <BoothMarker key={booth.id} booth={booth} scale={1} />
-  ))
-
   function updateBooths() {
     const feats = Array.from(featureMap.current.values())
     const booths = feats
@@ -179,16 +186,12 @@ export default function EditorMapComponent({
           feat.properties?.exhibitorId >= 0 // booths that were just created have exhibitorId -1
       )
       .map(feat => makeBooth(feat as GeoJsonBooth, exhibitorsByID))
-
     setBooths(booths)
   }
 
+  useEffect(initializeDraw, [mapLoaded])
+  useEffect(updateDrawFeatures, [location, mapLoaded])
   useEffect(updateBooths, [location, mapLoaded])
-
-  const activeFeature = activeFeatureId
-    ? featureMap.current.get(activeFeatureId)!
-    : null
-  const popupLocation = activeFeature ? getPolygonCenter(activeFeature) : null
 
   return (
     <div className="flex h-full w-full justify-center">
@@ -221,6 +224,8 @@ export default function EditorMapComponent({
             onClose={() => setActiveFeatureId(null)}>
             <FeatureEditor
               feat={activeFeature!}
+              setActiveFeatureId={setActiveFeatureId}
+              exhibitorsByID={exhibitorsByID}
               onSave={feat => {
                 setFeature(feat)
                 draw.add(feat)
@@ -237,136 +242,17 @@ export default function EditorMapComponent({
         )}
       </MapboxMap>
 
-      {mapLoaded && <Toolbar />}
+      {mapLoaded && (
+        <EditorToolbar
+          draw={draw}
+          drawMode={drawMode}
+          setDrawMode={setDrawMode}
+          resetFeatures={resetFeatures}
+          showBuildings={showBuildings}
+          setShowBuildings={setShowBuildings}
+          featureMap={featureMap}
+        />
+      )}
     </div>
   )
-
-  function FeatureEditor({
-    feat,
-    onSave,
-    onDelete
-  }: {
-    feat: GeoJsonBooth
-    onSave: (feat: GeoJsonBooth) => void
-    onDelete: (feat: GeoJsonBooth) => void
-  }) {
-    const props = feat.properties
-
-    const [exhibitorId, setExhibitorId] = useState<number>(props.exhibitorId)
-
-    return (
-      <div className="z-50 flex flex-col gap-4 divide-neutral-400 text-base text-stone-950">
-        <div className="">
-          <b>Booth ID:</b> {props.id}
-        </div>
-        <div className="flex items-center">
-          <b>Exhibitor:</b>
-          <select
-            className="ml-1 w-full rounded-md bg-stone-300 p-0.5 hover:cursor-pointer"
-            value={exhibitorId}
-            onChange={e => {
-              const newExhibitorId = parseInt(e.target.value)
-              setExhibitorId(newExhibitorId)
-              onSave({
-                ...feat,
-                properties: { ...props, exhibitorId: newExhibitorId }
-              })
-            }}>
-            <option value={-1}>None</option>
-            {Array.from(exhibitorsByID.values()).map(e => (
-              <option key={e.id} value={e.id}>
-                {e.name}
-              </option>
-            ))}
-          </select>
-        </div>
-
-        <div className="flex justify-between border-t border-stone-300 pt-3">
-          <button
-            className="w-max flex-none rounded-md bg-red-300 p-1 transition hover:bg-red-400"
-            onClick={() => {
-              onDelete(feat)
-            }}>
-            Delete
-          </button>
-          <button
-            className="w-max flex-none rounded-md bg-stone-300 p-1 transition hover:bg-stone-400"
-            onClick={() => {
-              setActiveFeatureId(null)
-            }}>
-            Close
-          </button>
-        </div>
-      </div>
-    )
-  }
-
-  function Toolbar() {
-    const [showJson, setShowJson] = useState(false)
-    const json = JSON.stringify(
-      toFeatureCollection(featureMap.current),
-      null,
-      2
-    )
-    const copyToClipboard = () => {
-      navigator.clipboard
-        .writeText(json)
-        .then(() => setDidCopy(true))
-        .catch(() => console.error("Failed to copy to clipboard"))
-    }
-    const [didCopy, setDidCopy] = useState(false)
-
-    return (
-      <div className="absolute bottom-2 flex gap-1">
-        <Button
-          className={cn(drawMode === "draw_polygon" && "dark:text-melon-700")}
-          onClick={() => {
-            if (!mapLoaded) return
-            draw.changeMode("draw_polygon")
-            setDrawMode("draw_polygon")
-          }}>
-          Draw
-        </Button>
-
-        <Button
-          onClick={() => {
-            setShowBuildings(!showBuildings)
-            resetFeatures(
-              showBuildings ? geoJsonBoothData : geoJsonBuildingData
-            )
-          }}>
-          {showBuildings ? "Edit booths" : "Edit buildings"}
-        </Button>
-
-        <div className="relative flex justify-center">
-          <Button onClick={() => setShowJson(!showJson)}>
-            {showJson ? "Hide JSON" : "Show JSON"}
-          </Button>
-          {showJson && (
-            <div className="absolute bottom-[45px] h-[500px] w-[500px] overflow-auto rounded-md bg-stone-200 p-2 text-stone-900">
-              {!didCopy ? (
-                <Button className="dark:bg-slate-400" onClick={copyToClipboard}>
-                  Copy to clipboard
-                </Button>
-              ) : (
-                <div>Copied!</div>
-              )}
-              <pre>{json}</pre>
-            </div>
-          )}
-        </div>
-
-        {/* reset button */}
-        <Button
-          className="dark:bg-gray-400"
-          onClick={() => {
-            resetFeatures(
-              showBuildings ? geoJsonBuildingData : geoJsonBoothData
-            )
-          }}>
-          Reset
-        </Button>
-      </div>
-    )
-  }
 }
